Updated Rating Structure based on new main.
